call windows replace function with MOVEFILE_WRITE_THROUGH flag
